requirements.txt
fastapi
uvicorn
jinja2
pillow
numpy
python-multipart
ultralytics
opencv-python-headless